Updated Build Setting + Important Cloning Note
for default PC resolution (this is a mobile game, but this is just in-case
Also, make sure that an appropiate resolution is set in the editor otherwise levels will look weird in their layouts. this means using one which mimics a typical phone aspect ratio. Or use the device simulator to mimic specific devices.